pyproject.toml
[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
name = "npserver"
version = "0.3"
requires-python = ">=3.11"

dependencies = [
    "arrow",
    "boto3",
    "click",
    "django>=5.0,<6.0",
    "django-admin-sortable2",
    "django-debug-toolbar",
    "django-eml-email-backend==0.1",
    "django-environ<1",
    "django-extensions",
    "django-prettyjson",
    "django-webpack-loader==0.6.0",
    "djangorestframework",
    "gunicorn",
    "loguru",
    "pginit",
    "psycopg2-binary",
    "pydantic[email]",
    "python-dateutil==2.8.1",
    "pytz",
    "redis==2.10.6",
    "requests<3",
    "sentry-sdk<3",
    "whitenoise[brotli]>=6.7.0",
    "Werkzeug<2",
    "typing-extensions",
    "django-select2",
    "django-anymail[mailgun,sendgrid]",
    "pendulum",
    "jinja2",
    "django-jinja",
    "ska>=1.10",
    "django-pydantic-field>=0.3.10",
    "django-ninja>=1.3.0",
    "httpx>=0.27.2",
    "pyjwt[crypto]>=2.9.0",
]
readme = "README.md"
license = {text = "MIT"}


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

# [tool.pdm]
# [[tool.pdm.source]]
# name = "pypi"
# url = "https://pypi.org/simple"
# verify_ssl = false


[tool.pdm.scripts]
_.env_file = "env-local"
uitest = "pytest -s -m uitest --headed --tracing retain-on-failure  --verbose"
test = "pytest -s -m 'not uitest' --doctest-modules -n auto --verbose"
doctest = "pytest -s -m 'not uitest' --doctest-modules -n auto --verbose ufo/"

# localization
pre_makemessages = "pybabel extract -F locale/babel.cfg -o ./locale/django.pot ./"
makemessages = "pybabel update -l ru -d locale/ -i locale/django.pot --domain django"
compilemessages = "pybabel compile -d locale/ --domain django"

[tool.pdm.dev-dependencies]
translate = [
    "Babel>=2.16.0",
]
test = [
    "pytest>=8.3.2",
    "beautifulsoup4",
    "pytest-django",
    "pytest-env",
    "pytest-pythonpath",
    "pytest-cov",
    "requests-toolbelt==0.8.0",
    "responses==0.10.12",
    "model-bakery>=1.19.5",
    "pytest-xdist>=3.6.1",
    "freezegun>=1.5.1",
]
debug = [
    "ipdb>=0.13.13",
]
uitest = [
    "playwright>=1.46.0",
    "pytest-playwright>=0.5.1",
]

